CosmosError: support usage of base class m_msg already during construction time
InternalError and UsageError unnecessarily hold an additional
std::string just to append it to the base class's std::string during
`generateMsg()` time.

Instead reuse the base class string object and change the `what()`
implementation to rely on a boolean flag to determine whether
`generateMsg()` needs to be called or not.