global: change include file structure in repository
Move all headers on level up outside of the include/cosmos directory.
Instead introduce a symlink cosmos to the include directory which allows
all include directives to work unchanged.

The advantage is that file paths in the repository become shorter and
are equal between the src and include directory. For system wide
installation still a proper include/cosmos directory can be used for
having a clean separation of libcosmos headers from other headers.