global: wiser use of const references for parameters
Use references only if the passed object is actually large or otherwise
requires the coupling to the passed object. Pass by value is the most
efficient method in C++ if the value is small, it also allows better
optimization on compiler level since there are no worries about
non-local changes to the values.