FileDescriptor: introduce strongly typed primitive FileNum
The primitive fd_t based on a typedef still allows things to go the
wrong way when mixed with something that is not actually a file
descriptor.

Therefore use a strongly typed primitive called `FileNum` here as well
to make it a proper type. Invoking system calls becomes a bit more
clumsy this way but the added safety and clarity should be worth it.