SubProc: split into ChildCloner and SubProc parts
Currently SubProc implements too much logic at once. It both prepares
and creates a child process but also represents the running state of a
created child process. With this change SubProc will merely remain as a
representation of the running child process, while the new ChildCloner
type contains all the logic to prepare and create a SubProc instance.

This change also harmonizes the ChildCloner API. There is now a
dedicated member to keep track of the executable to run, while argv[0]
is by default synchronized with it but can also be changed by users of
the class for special cases.

The SchedulerSettings are now stored by-value in the ChildCloner class.
This requires a somewhat complex std::variant to be used but reduces the
risk of dangling pointers.

The tracing option is dropped, since this tracing feature is better
implemented in a specialized TraceCloner type that can be created in the
future.

SubProc is now a non-copyable but movable type which allows to store a
SubProc instance e.g. in a member variable without having to worry about
ownership of the contained pidfd and the responsibility to wait() on the
child process.