fs/types.hxx: complete coverage of mode_t by introducing strong types
mode_t is a difficult field in file system related system calls. It
encodes both the file type and the permission bits and users a lot of
macros and bit magic.

This is an attempt in separating the two parts into strong types to
represent the file type (non-bitmask values) on the one hand and the
FileMode (permission and other file bits) on the other.