CosmosError: fix message content set at construction time (UsageError)
Somewhat a bug was introduced that causes message content already set at
construction time, like the case wit UsageError, to be left at the
beginning of the message, instead of at the end.

Fix this by remembering the original message content and appending it
after the base information has been applied.