global: another (hopefully the last) harmonization of coding style
This is mostly concerned with harmonizing spelling of function names and
type names:

- class functions always use camel style `doSomething()`
- global functions in namespaces use snake case `do_something()`. I'm
  a bit torn about this but it helps differentiating class context
  functions from global functions.
- type names start with a big Letter and camel case MyType throughout

Otherwise some smaller formatting changes and comment changes. And one
function rename for `linkChangeOwner()` to `change_owner_nofollow()`.
It's no longer obvious which system call this refers to, but it is much
more obvious what it does now.