time: introduce strongly typed TimeSpec coupled to the clock source
This greatly improves type safety for the TimeSpec type on a logical
level. Now only a TimeSpec from the proper clock source can be passed to
APIs. This removes the need for defining clock types in e.g. Condition
and PosixThread. It also makes it clear right from the TimeSpec
specialization which kind of time is used.

This moves TimeSpec.hxx to types.hxx and puts it together with the basic
ClockType. The ClockBase is now gone and we need explicit template
instantiations for each clock type.

Also provide typedefs and instantiations for all available clock types.