FileDescriptor: add support to duplicate with CLOEXEC without having to specify a valid new FD
dup3() actually only solves part of the race conditions involved with
file descriptors and O_CLOEXEC. Since it still requires a new_fd to
assign to, the caller needs to find an unused file descriptor in a race
free way, which isn't cleanly possible.

Via fcntl() this problem can be solves and this additional duplicate()
variant in FileDescriptor offers just that.