DirStream: change open(const DirFD) ownership semantics.
It is rather surprising that this open variant is taking ownership of
the file descriptor passed in. This is often only noticed once the
DirStream is closed and confusing.

To avoid this, duplicate the file descriptor and pass the duplicate into
fdopendir. That is another system call but worth the reduced surprise
factor.