BitMask: change semantics of test() and operator[]
The surprising aspects currently described is not really a good thing.
It seems the intuition is that when querying for an enum constant value
that *all* bits are checked, not just some.

The case that a mask with multiple bits is passed in and only for any of
the bits should be checked is a rather rare one.

Thus change the semantics, except for operator&, where a closer relation
to raw bit operations can be expected.

A current situation where the surprising effect comes up is when testing
OpenFlags against OpenSettings::TMPFILE, which is actually comprised
of two bits behind the scenes, O_TMPFILE and O_DIRECTORY. Thus a test
like `flags[OpenSettings::TMPFILE]` evaluates to true although only the
O_DIRECTORY bit is set in the mask. That is really confusing.