DirStream: support opening a DirStream directly relative to a DirFD
This allows to save a system call when recursively iterating over
directories. When opening a DirStream from an existing DirFD, then the
DirFD is currently duplicate() i.e. an fnctl() call has to happen.

If the original DirFD is not required though, then the DirStream can
constructed via an open_at() behind the scenes right away.