buildsystem: link static libs as object files, not using ´-l<lib>`
When switching between shared and static lib linking then currently
shared libraries will be prefered, due the usage of `-l<lib>` linking
even for static libraries.

To overcome this link static archives like object files, which is
generally the safer variant for static linking and allows to keep static
and shared artifacts around in the same build tree without worrying.

To support this the interface of configureForLib has to be changed,
since it can change the sources list of the library to be built.