global: harmonize naming scheme for BitMask and related enum classes
Currently the naming scheme for the various BitMask types are not very
intuitive. E.g. OpenSettings is the enum class type for individual bit
values, while OpenFlags is the BitMask type carrying a collection of
these bits.

For the FileMode context we have FileModeFlags denoting the enum class
type, while FileModeBits is the BitMask type carrying them.

To improve upon this follow a consistent naming scheme: The enum class
type is a singular term like `OpenFlag` and the BitMask type is a plural
term like `OpenFlags`. Since both are strong types there is no danger
mixing them up.