global: replace include guards by #pragma once
To reduce noise in headers and also the chance for errors, replace
traditional include guards by #pragma once.

While #pragma once is non-standard is is supported by the majority
compilers. It should also give us a little edge on compilation times,
since less files need to be opened and processed by the compiler.