main.hxx: fix static_assert in else{} branch on clang
Compiling on clang breaks for main.hxx, because the `static_assert` is
triggered (false positive). The reason is that formally a constexpr if
branch is not compiled only if it is located in a template, and for
static_assert it is only not evaluated if it depends on a template
parameter.

`static_assert(false, ...)` can always be evaluated and this actually
happens on clang. To get the desired result we need to depend on a
template parameter but find a condition that will always fail if the
branch is actually fully instantiated.

Doing it this way seems to work.