ChildCloner: introduce transparent fallback for fork() in Valgrind context
Not being able to run programs built on top of libcosmos in Valgrind
when ChildCloner is involved is unfortunated. Dropping the clean clone3
API just because of this would also be bad. Thus try to make both work:
Keep the clone3() approach for regular operation but fall back to a more
complex and error prone fork() approach, if a Valgrind runtime
environment is detected.