Generalize valgrind dtection concept and add fallback for pidfd_send_signal
It turns out there are more system calls that cause problems when
running in a Valgrind environment. Thus change the semantics of
prefer_clone to a general valgrind runtime detection flag.

Based on this flag also switch the implementation of signal::send(const
PidFD, const Signal).

Also document this larger scale of the problem in the README.