FileDescriptor: fix fcntl(F_DUPFD_CLOEXEC) API
It turns out this `fcntl()` requires an `arg` to denote the file
descriptor number to start with. This mostly worked out (by returning
some random file descriptor number) but sometimes an EINVAL is the
result if the value is too large.

Fix this by extending the client API and passing the arg as required.