PosixThread: fix race condition in setting m_pthread
If a thread is started that gets access to the PosixThread object then a
race condition is currently involved, because the optional `m_pthread`
may not yet be assigned while the thread is already alive.

To fix this assign an invalud value to m_pthread and let
`pthread_create` overwrite it before the thread is created.

This is not ideal modelling wise but other solutions would require
setting up some form of explicit synchronization for this rather rare
use case.