global: use angle bracket includes for libcosmos
Given that we use a hierarchical include structure the preference for
the local directory is not ideal.

Furthermore due to the symlink we currently have in include/cosmos, this
leads to problematic source paths being embedded in resulting ELF
binaries like cosmos/cosmos/cosmos/something.hxx. This confuses e.g.
generation of -debugsource packages in RPM packaging.

The disadvantages of using angle brackets here should be negligible.