.github: add github action workflow file
This will build and run unit tests on the latest ubuntu runner on GitHub
in all configurations that work.

GitHub's Ubuntu image is in some ways non-standard and e.g. doesn't
properly provide 32-bit mulitlib packages and also static linking with
clang is for some reason broken. Reproducing this locally is hard, since
the image needs to be built using complex image building tooling.

So I decide simply to skip these problematic configurations for now.
They can be tested locally by running check.py.

Also Doxygen does not work as expected in the Ubuntu image. I did not
check yet as of why (the -q switch seems to be unsupported there).

On the plus side we can also test an arm cross build.