buildsystem: lookup pkg-config templates in current project_root
When libcosmos is part of a sub-module then it will currently look in
the top level of the project for a matching input file. This won't work
out. Asking `git` where our current top-level is would be nice, but this
also needs to work in a source tarball.

Thus mark the current project root in the `env` dictionary for looking
up the templates.