SubProc: fix object state when wait() is called on a killed process
Currently when a child process is killed by a signal and this
information is collected via SubProc::wait(), then the SubProc's state
is not properly reset. This then leads to follow-up errors, when
destroying the SubProc instance, or other logical errors, because
SubProc::running() still returns `true`.

Also add a basic test to the unit tests that catches this issue.