signal headers: split part of Signal.hxx into signal.hxx
This is to avoid circular include dependencies. The Signal type should
be available without a lot of dependencies to things like the PidFD. The
more complicated stuff and free functions are now placed into
signal.hxx. It's a bit unfortunate that the headers only differ in case,
but it matches the coding style: Signal.hxx contains the class type
while signal.hxx contains free functions and constants.