sigqueue wrapper: use std::variant instead intptr_t
We cannot assume that processes we're sending signals to use libcosmos.
The target processes might only interpret the `si_int` field, which can
be smaller than `si_ptr`. This would make it hard for users of libcosmos
to send the proper data using the wrapper. Thus use a std::variant
instead to make it possible to send out either type.