SignalFD: replace SignalFD::SigInfo by new structure based on SigInfo
signalfd_siginfo_t and siginfo_t are very similar, but equal. Thus it
would lead to awkward and complex code when trying to merge them. This
means we have to swallow the bitter pill that parts of the SigInfo
implementation have to be copied for the new SignalFD::SigInfo type.

With this change the SignalFD provides full data access to signal data.