signal: add pseudo SYS_TRAP for use with ptrace(2)
There can be special SIGTRAP reports with siginfo_t when
PTRACE_O_TRACESYSGOOD is involved. The SIGTRAP value is or'ed with 0x80
in this case. Such kind of bit operations are awkward to work with when
otherwise strong enums are used in libcosmos. Thus add this as a pseudo
signal to check for when dealing with ptrace().