SubProc: fix object state for wait() with WaitFlag::STOPPED
There is another inconsistency in SubProc if WaitFlag::STOPPED is
specified and the child process actually enters the STOPPED state. In
this case SubProc will wrongly be invalidated, although it is still
alive.

Add a test case to catch this situation and fix the logic by only
invalidating object state if the child exited or got killed.