ptrace API: exclude more ptrace::Request constants if not available
These constants are not available on aarch64. I don't see a better way
than to use the preprocessor to exclude them and also the wrappers in
Tracee at the moment.

Users of these requests likely are pretty low-level themselves, so
dealing with some #ifdef's will hopefully not be too much of a burden on
them.

Fixes #1