ChildData: split type into ChildState for wait() and ChildData for SigInfo
ChildData depending SigInfo is problematic with regards to the ability
to forward declare the type. It is impossible to forward declare nested
classes.

The proc::wait() functions have no logical relationship to SigInfo and
SigInfo is also a complex header. So requiring users of proc::wait() to
include it is not good.

Also ChildData has fields that are not used in the proc::wait() context.
Thus split the type in two: The new ChildState contains the bulk of the
data and is now a basic type in proc/types.hxx. It is used in the
proc::wait() functions.

ChildData remains a nested type in SigInfo but now inherits from
ChildState to share its fields and logic.

Adjust all users of ChildData accordingly.