BitMask: support implicit construction from enum values
To make bit mask usage less verbose support creating BitMask from the
enum values in an implicit fashion. Since the enum class and the BitMask
are usually closely tied in usage and the enum class is a strong type on
its own, there shouldn't be much danger in allowing this.