exception system: use std::source_location instead of macros
With C++20 we can finally use std::source_location as a builtin way to
manage exception source location. With this things become a lot nicer:

- the exception object actually becomes smaller than before
- we no longer need the cosmos_throw macro to fill in the source
  location info
- thus we also no longer need the raise() method and what is involved
  with it

We no longer get bogus compiler warning due to the compiler not
understanding that cosmos_throw does not return. We can use curly braces
initialization syntax for exceptions.