process.hxx: exec family of functions: more sensible string vector handling
The mixture of references and pointers here is confusing. Some variants
have been incomplete. Turn the conversion functions to CStringVector
into inline functions.

Using std::optional<> for args and env would be nice, but it would
require a copy when clients want to reuse a fixed object for arguments
or environment variabless. So stick to using references and a bunch of
variants of function signatures.