BitMask: add steal() member function
This new member comes in handy when you want to test for a bit, and if
it's set, run some specific code and reset the bit right away. So you
can do something like:

    if (mask.steal(Flag::ONE_SHOT)) {
        oneShot();
    }

And the Flag::ONE_SHOT will be reset without adding more noise or
possibility for error.